Use config to determine how long uproebstats runs for
Task config has a duration_seconds field that was previously unused.
This change updates UprobeStats to use that field to determine how long
to run for.

At the moment we only plan to support single-task configs, so this
change also logs an error if anything other than such a config is found,
and only looks at the first entry if so.

Bug: 296108553
Test: hello_uprobestats (and trigger something that calls `Slog` - log
output should show the call being detected)

Change-Id: I7326b20facace676f7ea080cfacc7ea15da19ad6