Add ability for UprobeStats to write to StatsD
This change adds a field `primitive_argument_positions` to the
`statsd_logging_config` field of the proto configuration for
UprobeStats.

If the config is present, and the generic BPF is used, UProbeStats pulls
the values returned from the BPF accordingly.  The integers passed
should represent the argument positions in a probed java method, and
only primitive types (e.g. int) should be used.

Bug: 296108553
Test: hello_uprobestats -n test_bss_noteEvent
Flag: NONE UprobeStats doesn't yet run on devices
Change-Id: Ibac602c5b54ac928d8e66f20a840885f8c8bcc34