Adds call timestamp bpf and updates smoke test
Mainly for testing purposes, add a generic bpf that simply records a
timestamp when the probed function is called. This is easier to test
cross-platform (arm/x86).

Update the main uprobestats program to write this timestamp to an atom
if provided in the config.

Update the smoke test to use this bpf, and enable it in postsubmit by
adding TEST_MAPPING.

Bug: 296108553
Test: atest uprobestats-test
Flag: NONE UprobeStats does not yet run on devices
Change-Id: Ice1ac8c65bd225c35fde2e010fdd8cea4355d7fa