Conditionally use flags in mainline container
uprobestats is in the progress of being moved from platform to mainline.
To ensure it's using the flags from the right container across various
build configurations, make it conditionally use flags from either the
"system" or the "mainline" container depending on whether it is already
in Mainline.

Bug: 355315669
Test: TH
Flag: build.RELEASE_UPROBESTATS_MODULE
Change-Id: I20771745b51d14ac7074f343f34786e543b4a0bc