Bare minimum bitmap instrumentation in rust
In addition to a new BitmapAllocation BPF as well as glue code to
process its results, this change introduces some nontrivial changes that
allow us to fetch the pid and uid of a process once it starts.

It adds new fields to the config proto for a client to specify that they
want instrumentation to run as soon as any process starts, or as soon as
a particular app (identified by package name) process starts.

This also includes some housekeeping changes to isolate tests that
should only run in the rust implementation of uprobestats.

Bug: 400457896
Bug: 409775699
Test: atest uprobestats-test:test.SmokeTestRustOnly#bitmapAllocation
Flag: android.uprobestats.mainline.flags.enable_bitmap_instrumentation
Change-Id: I8867ecf195b66950f08e0a1529ae2fce7d9b1137