Read log level from log.tag.uprobestats
Sets the uprobestats log level based on a system property, defaulting to
INFO if it is not set or fails to parse.

On user build, the max allowed level is still INFO (but the property can
be used to make it lower).

Bug: 408477186
Test: atest uprobestats-test
Flag: EXEMPT uprobestats_rs not yet executing on devices
Change-Id: I647dcae97e66c558ba1cc24d693afea7f666315b