Small fixes to ramp RELEASE_UPROBESTATS_RUST
Fix oatdump fallback:

When the platform API to fetch method offsets is unavailable, or the
delivered config doesn't support it, we fallback to oatdump. This logic
was failing emulator tests on x86 because we weren't handling multiple
paths for the odex file in config (bailing when the first one wasn't
found). This code path *shouldn't* execute on real devices but a fix is
made since we *do* test it, and it's testable on x86.

Fail fast on unknown BPFs in legacy implementation:

There was an `else` clause in the legacy implementation that would just
print some debug info when an unknown BPF map was encountered. Instead -
explicitly exit - as a performance optimization in the case those
devices receive configs intended for the new implementation.

This change also adds an arm64 assumption on tests that will only pass
on arm64 devices.

Bug: 408259024
Test: uprobeestats-test
Flag: build.RELEASE_UPROBESTATS_RUST
Change-Id: Ic4b643be061bf1c3331ebd09eac711bbe318880e