Allow stateful BPF map handlers
Currently, the BPF map handlers are stateless. This commit changes them
so that they can hold state, and this state is preserved across multiple
calls to `on_item`. This is important for implementing various rate
limiting mechanisms, where handlers need to be able to "remember" the
previous calls.

This is achieved by:
*  Changing the trait `OnItem` to `Handler`, and passing mutable reference to `on_item` method
* Adding a default implementation to handlers, which create a default mutable object at the begin of the BPF map polling loops.

Bug: 400457896
Test: m com.google.android.uprobestats && adb install out/soong/.intermediates/packages/modules/UprobeStats/apex/com.android.uprobestats/android_common_com.google.android.uprobestats_com.google.android.uprobestats/com.google.android.uprobestats.apex && adb reboot && atest uprobestats-test
Flag: EXEMPT refactor
Change-Id: Iab9dbc8cbfc3ee11a4577df80cf30f1e70c5e703