Introduce BpfMapAccessor for read/write from userspace
Previous to this change, bpf maps (and ring buffers) defined in the
module were only written to from inside an eBPF programs. Userspace only
ever *read* from *ring buffers*.

Incoming use cases will require to support reading from and writing to
bpf maps in userspace.

Introduce a abstraction to do this, to be used in a subsequent change.

Bug: 435672760
Test: TH
Flag: EXEMPT abstraction is unused in this commit
Change-Id: I0ec2fd013a5b32d456c931362fde4dc9ca618296